+++ src/script/constants/constants.js
@@ -1,0 +1,23 @@
+// Keyboard Keys
+export const keys = {
+    ENTER: 'Enter',
+    TAB: 'Tab',
+    ESCAPE: 'Escape',
+    ARROW_LEFT: 'ArrowLeft',
+    ARROW_RIGHT: 'ArrowRight',
+};
+
+// Events
+export const events = {
+    CLICK: 'click',
+    KEYDOWN: 'keydown',
+    SCROLL: 'scroll',
+};
+
+// Attributes
+export const attributes = {
+    ARIA_LABEL: 'aria-label',
+    ARIA_EXPANDED: 'aria-expanded',
+    ARIA_HIDDEN: 'aria-hidden',
+    INERT: 'inert',
+};
